push gql fetching and streaming into topsymbols
