+++ src/components/YuzuHeader/YuzuHeader.tsx
@@ -4,7 +4,7 @@ const YuzuHeader = () => {
   return (
     <div className="navigation flex-shrink fixed w-full">
       <nav>
-        <Link to="finance">
+        <Link to="/">
           <h2>
             🍋 <span className="">Yuzu</span>{" "}
             <span className="text-amber-600 font-sans font-normal text-xl">
